shell/requirements.txt
Package                  Version
------------------------ -----------
aiohappyeyeballs         2.6.1
aiohttp                  3.12.4
aiosignal                1.3.2
asttokens                3.0.0
attrs                    25.3.0
certifi                  2025.4.26
charset-normalizer       3.4.2
click                    8.1.8
dbus-python              1.4.0
evdev                    1.9.2
executing                2.2.0
fabric                   0.0.2
frozenlist               1.6.0
holidays                 0.73
idna                     3.10
iniconfig                2.1.0
jeepney                  0.9.0
littleutils              0.2.4
loguru                   0.6.0
material-color-utilities 0.2.6
multidict                6.4.4
numpy                    2.3.0
nvidia-ml-py             12.575.51
nvidia-smi               0.1.3
packaging                25.0
pillow                   12.0.0
pip                      25.0.1
pip-check-reqs           2.5.5
pluggy                   1.6.0
propcache                0.3.1
psutil                   7.0.0
pulsectl                 24.12.0
pycairo                  1.28.0
Pygments                 2.19.1
PyGObject                3.50.0
pynput                   1.8.1
pynvml                   12.0.0
PySensors                0.0.4
pytest                   8.4.0
python-dateutil          2.9.0.post0
python-xlib              0.33
requests                 2.32.3
requirements-parser      0.13.0
ruff                     0.11.12
setproctitle             1.3.6
six                      1.17.0
sorcery                  0.2.2
urllib3                  2.4.0
wrapt                    1.17.2
yarl                     1.20.0
